Changes list priority from int to double.
This allows an easy implementation of something like A* in work queue,
while preserving the old int behaviour.

+++ dttools/src/list.c
@@ -133,7 +133,7 @@ static struct list_node *new_node(void *data, struct list_node *prev, struct lis
 	return node;
 }
 
-int list_push_priority(struct list *l, void *item, int priority)
+int list_push_priority(struct list *l, void *item, double priority)
 {
 	struct list_node *n;
 	struct list_node *node;
